[TASK] Improved t3jquery integration
The JavaScript file that should be included can now be configured with
a constant.

Additionally the inclusion method was changed so that the t3jquery setting
"integrateToFooter" will be respected.

Finally the manual was improved and now contains a hint that the static
TypoScript template "Image tooltip rendering" is still used.

Thanks to Harald Witt for test feedback.

Relates: #51968